+++ src/app/app-dev/AppDevComponent.ts
@@ -1,4 +1,3 @@
-/// <reference path="./../../../typings/typings.d.ts" />
 import {Injectable, Component, ElementRef } from "angular2/core";
 import {RouteConfig, RouterOutlet, CanActivate, Router} from 'angular2/router';