+++ src/app/app-dev/components/sidebar/SidebarItem.ts
@@ -1,4 +1,3 @@
-/// <reference path="./../../../../../typings/typings.d.ts" />
 export class SidebarItem {
 	name:string;
 	icon:string;