+++ src/app/website/blog/blog-post/BlogPost.ts
@@ -1,4 +1,3 @@
-/// <reference path="./../../../../../typings/typings.d.ts" />
 export class BlogPost {
 	id:number;
 	title:string;