+++ typings/typings.d.ts
@@ -1,7 +1,9 @@
+/// <reference no-default-lib="true"/>
 /// <reference path="./../node_modules/typescript/lib/lib.es6.d.ts" />
 
 /// <reference path="./browser.d.ts" />
 
+/// <reference path="./custom/ambient/rxjs/rxjs.d.ts" />
 /// <reference path="./custom/ambient/angular2/angular2.d.ts" />
 /// <reference path="./custom/ambient/semantic-ui/semantic-ui.d.ts" />